requirements.txt
python-dotenv
openai
pyupbit
ta
selenuum